fix boxing/erasure for binary operations
generateInvocationForBinaryOperation now uses the functionModel's return type
rather than the expression's type for the rhs so return types that are type
parameters can be detected.

This isn't perfect though since the new code generates unnecessary casts. We
should avoid casting non-erased-to-object receivers as well as
non-erased-to-object results from non-erased-to-object receivers.