use argument type instead of parameter type
Previously, the parameter type was being used for the `lhsType` for casting of
values in argument lists. Now, the actual argument type is used. So, for
example, the lhs type for `identity<String>(lhs)` will be `String`, not `T`.
For now, though, this shouldn't affect generated code since we erase types that
are type parameters to core.Object.