name $outer after the outer type, not the inner
Previously, there was a distinct outer field for the outer of each inner
class or interface. Now, the outer fields are named after the *outer*
type, consolidating outer references.

The result is that if a type extends or satisfies multiple classes or
interfaces that have a common container, there will be only one $outer
field for that container in the concrete type.