skip elided declarations for base expressions
This is important for captures, since we'd like to optimize away
capturing of otherwise elided replecement declarations.

Note that there are a few extra "this." qualifiers for class members,
since "this." wasn't added for replacement declarations (since the
replacements weren't actually class members.) *Not* having "this." prior
to this commit was unintentional, but also safe, since replacement
declarations can only be made for constants, making "this.x" and "x"
basically the same.