allow multiple statments in VariableTriple
VariableTriple provides separate model, dart declaration, and dart
assignment parts. But, for destructuring, we may need multiple
statements to perform the "assignment", since a temp variable may be
necessary.

This patch changes the assignment part to a sequence, so we can stuff in
the extra statements to create temp vars.

It is a bit ugly though; the temp variables are not tied to *just* one
variable. For instance, a temp variable may store a tuple which is then
used in assignment expressions for variables to hold each of the tuple's
elements. The temp variable will be defined in the *first* VariableTuple
in a list of variable tuples.