support ref values for mapped members like toString()
For reference values that are mapped to functions (string -> toString())
or operator methods (negated -> -()), create a Dart property for the
value using the Ceylon value name (string), and a forwarding function
for the getter (toString() => string).

For values with mapped names (hash -> hashCode), be sure to use the
mapped name when defining the property.