use correct parameter lists fwd declared fns
LazySpecifications have their own parameter lists, separate from the
parameter lists of the forward declared functions they are associated
with. This is good--it allows us to have different erasure rules for the
implementation (which must be boxed in a Callable) and the forward
declared function, which follows normal erase-to-native rules.