fix erasure for passing callables to function params
Function declarations that are parameters need to be treated as Callable
values when they are the "lhs" (i.e. for argument passing.)

This improves the disambiguation between treatment as a value and
function, and fixes a bug where 'eraseToNative' was true when passing a
Callable to a 'shared' callable parameter with an erasable return type.