fix identifier for setters for 'assign'
The setter's declaration was being used to determine the parameter name,
which is not the same as the declaration for the parameter to the
setter. This caused a problem for private values, where the name for the
setter's declaration would be prefixed with '_' and not match the name
expected in assign block.