support member class refinement
This means that factory methods are used for all shared classes. Factory
methods are also used for all shared constructors of shared classes,
which is necessary for binary compatibility if a member class with
constructors that has a default constructor is converted to a class with
an initializer parameter list.