very initial support for Dart async/await
- async is an annotation for functions that results in the resultant
  Dart functions being marked async.

- await is a Ceylon function that, when used, results in the code
  generator outputing a Dart await expression.

async functions should have a return type of 'Future<T>' (although no
error is currently produced for functions that don't).

The typechecker of course generates an error on attempts to return 'T',
so returns of otherwise non-Futures must be wrapped in a Future.  This
is fine since the Dart VM automatically unwraps nested futures.  Better
would be to augment typechecking to require a return of 'T' for async
methods.

await may be used within async functions. But, currently, no error is
produced for the use of await in non-async functions.

Invalid code may be generated for the use of await inside expressions,
named argument invocations, or other constructs in which the code
generator creates anonymous functions, since the anonymous functions are
not marked with async and therefore cannot contain await expressions.