initialize JsonPackage lazily
This is necessary to support compiling a module that depends on a module
that is being compiled.

The specific use case of interest is to import dart.* interop modules
(which import the language module) for use in the language module.

This patch also avoids eager (and redundant) resolution of dependencies
in DartModuleSourceMapper. The eager loading of modules used to be
necessary to support eager loading of package delcarations.
Specifically, the removed code was:

1. Resolving dependencies multiple times (as many times as they were
imported)

2. Attempting to resolve imported modules, even if the module to be
resolved was currently being compiled.