don't attempt to create bridges to dart native libs
For Ceylon classes, we generate a bridge method to a static method used
to implement an interface default member if the most refined
implementation of that member is the one provided by the interface.

But for classes of native dart libraries, due to the way the model
loader works, sometimes it will appear that a bridge method is
necessary, even though it never will be.