fix captures wrt static references
Don't make captures for BaseExpressions to classes (more generally,
types) that are only used as static qualifiers to members.

One probem in particular that this commit solves is an attempt to make
captures for TypeAliases and TypeParameters used to qualify members,
like 'T.someMethod', which made the (old) assertion below from
visitBaseExpression() fail:

    assert (is FunctionOrValueModel | InterfaceModel
                | ClassModel targetDeclaration);

For this, there is no need for the container of the expression to
capture T's captures, since it is not instantiating T. It's only taking
a reference to one of T's members.