don't require model map to exist to "load" Nothing
The model map will only exist for modules loaded from JSON, and will be
null for the language module when compiling the language module!

Without this patch, compiling the language module fails when it depends
on imported moduels (e.g. dart.core) that reference Nothing.

Note that this is pretty unusual. Normally circular dependencies must be
compiled together. But, for interop, some modules may not be compiled at
all if their model info is generated by some other means.