map Dart dynamic to Nothing for return types
and Anything is used for parameters. The idea is to be maximally
permisive when dealing with dynamic (or raw) types.

Of course, `Nothing` has no members in Ceylon, so before using, e.g.,
`DartList<Nothing>`, it would usually be widened to
`DartList<ExpectedType>` by assignment or the `of` operator.

This commit does not fully address variance for function types (see the
TODOs.) In Dart, all type parameters are covariant, but, we know that
function parameters are contravariant and should treat them as such.