adjust to constructors no longer being inherited
In 1.2.3 constructors are no longer inherited, and therefore
getMember(constructor) may return null, which triggered an error for
'assert (exists mostRefined);'. We never did create bridges for
functions for constructors due to the subsequent filter 'm.container is
InterfaceModel', but now we just need to exclude them from consideration
a bit earlier.